requirements.txt
ansicolors==1.0.2
boto3==1.4.0
botocore==1.4.49
cffi==1.6.0
configparser==3.5.0b2
cryptography==1.3.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.5
django-filepicker==0.2.2
django-gulp==2.1.1
django-rest-swagger==0.3.5
django-watson==1.2.1
djangorestframework==3.3.3
djoser==0.4.0
docutils==0.12
env-tools==2.0.0
futures==2.2.0
gunicorn==19.4.5
idna==2.1
jmespath==0.9.0
Markdown==2.6.4
paypalrestsdk==1.11.5
psutil==2.2.1
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
pystache==0.5.4
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
requests==2.9.1
s3transfer==0.1.2
six==1.10.0
static==1.1.1
static3==0.6.1
tini==3.0.1
whitenoise==2.0.6
